backend/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.6.0
django-encrypted-model-fields==0.6.5
python-decouple==3.8
psycopg2-binary==2.9.10
cryptography==43.0.3
